Camera-event-recognition/requirements.txt
flask_restplus==0.13.0
flask-caching==1.9.0
flask-cors==3.0.9
Flask==1.1.2
numpy~=1.19.4
opencv_python==4.4.0.44
werkzeug==0.16.1

opencv-python~=4.4.0.46
scipy~=1.5.4
Pillow~=8.0.1